Updated to Laravel 7.3

+++ database/migrations/2018_11_09_135510_update_table_users_add_column_department.php
@@ -14,7 +14,7 @@ class UpdateTableUsersAddColumnDepartment extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->integer('department_id')->unsigned()->after('is_admin')->default(1);
+            $table->bigInteger('department_id')->unsigned()->after('is_admin')->default(1);
 
             $table->foreign('department_id', 'user_fk_department')->references('id')->on('departments');
         });
